fix: remove legacy prefix from vue2 components
