+++ legacy/sandboxes/bar/src/components/Bar.vue
@@ -1,5 +1,5 @@
 <template>
-  <Pie
+  <Bar
     :chart-options="chartOptions"
     :chart-data="chartData"
     :chart-id="chartId"
@@ -13,28 +13,29 @@
 </template>
 
 <script>
-import { Pie } from 'vue-chartjs/legacy'
+import { Bar } from 'vue-chartjs/legacy'
 
 import {
   Chart as ChartJS,
   Title,
   Tooltip,
   Legend,
-  ArcElement,
-  CategoryScale
+  BarElement,
+  CategoryScale,
+  LinearScale
 } from 'chart.js'
 
-ChartJS.register(Title, Tooltip, Legend, ArcElement, CategoryScale)
+ChartJS.register(Title, Tooltip, Legend, BarElement, CategoryScale, LinearScale)
 
 export default {
-  name: 'LegacyPie',
+  name: 'BarChart',
   components: {
-    Pie
+    Bar
   },
   props: {
     chartId: {
       type: String,
-      default: 'pie-chart'
+      default: 'bar-chart'
     },
     datasetIdKey: {
       type: String,
@@ -64,11 +65,25 @@ export default {
   data() {
     return {
       chartData: {
-        labels: ['VueJs', 'EmberJs', 'ReactJs', 'AngularJs'],
+        labels: [
+          'January',
+          'February',
+          'March',
+          'April',
+          'May',
+          'June',
+          'July',
+          'August',
+          'September',
+          'October',
+          'November',
+          'December'
+        ],
         datasets: [
           {
-            backgroundColor: ['#41B883', '#E46651', '#00D8FF', '#DD1B16'],
-            data: [40, 20, 80, 10]
+            label: 'Data One',
+            backgroundColor: '#f87979',
+            data: [40, 20, 12, 39, 10, 40, 39, 80, 40, 20, 12, 11]
           }
         ]
       },
