+++ legacy/test/Line.spec.js
@@ -1,19 +1,19 @@
 import { mount } from '@vue/test-utils'
-import LegacyBar from './examples/LegacyBar.vue'
+import LegacyLine from './examples/Line.vue'
 
-describe('LegacyBar', () => {
+describe('LegacyLine', () => {
   const Component = {
-    template: '<div><LegacyBar :chart-id="chartId" :plugins="plugins" /></div>',
-    components: { LegacyBar },
+    template: '<div><LegacyLine :chartId="chartId" :plugins="plugins" /></div>',
+    components: { LegacyLine },
     props: ['chartId', 'plugins']
   }
 
   it('should render a canvas', () => {
     const wrapper = mount(Component)
 
-    const barChart = wrapper.find('#bar-chart')
-    expect(barChart.element.id).not.toBe('undefined')
-    expect(barChart.exists()).toBe(true)
+    const lineChartEl = wrapper.find('#line-chart')
+    expect(lineChartEl.element.id).not.toBe('undefined')
+    expect(lineChartEl.exists()).toBe(true)
 
     const canvasEl = wrapper.find('canvas')
     expect(canvasEl.exists()).toBe(true)
@@ -21,12 +21,12 @@ describe('LegacyBar', () => {
 
   it('should change id based on prop', () => {
     const wrapper = mount(Component, {
-      propsData: { chartId: 'barchartprop' }
+      propsData: { chartId: 'linechartprop' }
     })
 
-    const barChart = wrapper.find('#barchartprop')
-    expect(barChart.element.id).not.toBe('undefined')
-    expect(barChart.exists()).toBe(true)
+    const lineChartEl = wrapper.find('#linechartprop')
+    expect(lineChartEl.element.id).not.toBe('undefined')
+    expect(lineChartEl.exists()).toBe(true)
   })
 
   it('should add inline plugins based on prop', () => {
