+++ legacy/test/Pie.spec.js
@@ -1,19 +1,19 @@
 import { mount } from '@vue/test-utils'
-import LegacyLine from './examples/LegacyLine.vue'
+import LegacyPie from './examples/Pie.vue'
 
-describe('LegacyLine', () => {
+describe('LegacyPie', () => {
   const Component = {
-    template: '<div><LegacyLine :chartId="chartId" :plugins="plugins" /></div>',
-    components: { LegacyLine },
+    template: '<div><LegacyPie :chartId="chartId" :plugins="plugins" /></div>',
+    components: { LegacyPie },
     props: ['chartId', 'plugins']
   }
 
   it('should render a canvas', () => {
     const wrapper = mount(Component)
 
-    const lineChartEl = wrapper.find('#line-chart')
-    expect(lineChartEl.element.id).not.toBe('undefined')
-    expect(lineChartEl.exists()).toBe(true)
+    const pieChartEl = wrapper.find('#pie-chart')
+    expect(pieChartEl.element.id).not.toBe('undefined')
+    expect(pieChartEl.exists()).toBe(true)
 
     const canvasEl = wrapper.find('canvas')
     expect(canvasEl.exists()).toBe(true)
@@ -21,12 +21,12 @@ describe('LegacyLine', () => {
 
   it('should change id based on prop', () => {
     const wrapper = mount(Component, {
-      propsData: { chartId: 'linechartprop' }
+      propsData: { chartId: 'piechartprop' }
     })
 
-    const lineChartEl = wrapper.find('#linechartprop')
-    expect(lineChartEl.element.id).not.toBe('undefined')
-    expect(lineChartEl.exists()).toBe(true)
+    const pieChartEl = wrapper.find('#piechartprop')
+    expect(pieChartEl.element.id).not.toBe('undefined')
+    expect(pieChartEl.exists()).toBe(true)
   })
 
   it('should add inline plugins based on prop', () => {
