+++ legacy/test/Scatter.spec.js
@@ -1,20 +1,20 @@
 import { mount } from '@vue/test-utils'
-import LegacyBubble from './examples/LegacyBubble.vue'
+import LegacyScatter from './examples/Scatter.vue'
 
-describe('LegacyBubble', () => {
+describe('LegacyScatter', () => {
   const Component = {
     template:
-      '<div><LegacyBubble :chartId="chartId" :plugins="plugins" /></div>',
-    components: { LegacyBubble },
+      '<div><LegacyScatter :chartId="chartId" :plugins="plugins" /></div>',
+    components: { LegacyScatter },
     props: ['chartId', 'plugins']
   }
 
   it('should render a canvas', () => {
     const wrapper = mount(Component)
 
-    const bubbleChartEl = wrapper.find('#bubble-chart')
-    expect(bubbleChartEl.element.id).not.toBe('undefined')
-    expect(bubbleChartEl.exists()).toBe(true)
+    const scatterChartEl = wrapper.find('#scatter-chart')
+    expect(scatterChartEl.element.id).not.toBe('undefined')
+    expect(scatterChartEl.exists()).toBe(true)
 
     const canvasEl = wrapper.find('canvas')
     expect(canvasEl.exists()).toBe(true)
@@ -22,12 +22,12 @@ describe('LegacyBubble', () => {
 
   it('should change id based on prop', () => {
     const wrapper = mount(Component, {
-      propsData: { chartId: 'bubblechartprop' }
+      propsData: { chartId: 'scatterchartprop' }
     })
 
-    const bubbleChartEl = wrapper.find('#bubblechartprop')
-    expect(bubbleChartEl.element.id).not.toBe('undefined')
-    expect(bubbleChartEl.exists()).toBe(true)
+    const scatterChartEl = wrapper.find('#scatterchartprop')
+    expect(scatterChartEl.element.id).not.toBe('undefined')
+    expect(scatterChartEl.exists()).toBe(true)
   })
 
   it('should add inline plugins based on prop', () => {
