+++ legacy/test/examples/Line.vue
@@ -1,5 +1,5 @@
 <template>
-  <Doughnut
+  <LineChartGenerator
     :chart-options="chartOptions"
     :chart-data="chartData"
     :chart-id="chartId"
@@ -13,28 +13,28 @@
 </template>
 
 <script>
-import { Doughnut } from '../../src/index'
+import { Line as LineChartGenerator } from '../../src/index'
 
 import {
   Chart as ChartJS,
   Title,
   Tooltip,
   Legend,
-  ArcElement,
-  CategoryScale
+  LineElement,
+  LinearScale
 } from 'chart.js'
 
-ChartJS.register(Title, Tooltip, Legend, ArcElement, CategoryScale)
+ChartJS.register(Title, Tooltip, Legend, LineElement, LinearScale)
 
 export default {
-  name: 'LegacyDoughnut',
+  name: 'LineChart',
   components: {
-    Doughnut
+    LineChartGenerator
   },
   props: {
     chartId: {
       type: String,
-      default: 'doughnut-chart'
+      default: 'line-chart'
     },
     datasetIdKey: {
       type: String,
@@ -64,11 +64,20 @@ export default {
   data() {
     return {
       chartData: {
-        labels: ['VueJs', 'EmberJs', 'ReactJs', 'AngularJs'],
+        labels: [
+          'January',
+          'February',
+          'March',
+          'April',
+          'May',
+          'June',
+          'July'
+        ],
         datasets: [
           {
-            backgroundColor: ['#41B883', '#E46651', '#00D8FF', '#DD1B16'],
-            data: [40, 20, 80, 10]
+            label: 'Data One',
+            backgroundColor: '#f87979',
+            data: [40, 39, 10, 40, 39, 80, 40]
           }
         ]
       },
