+++ legacy/test/examples/Radar.vue
@@ -1,5 +1,5 @@
 <template>
-  <PolarArea
+  <Radar
     :chart-options="chartOptions"
     :chart-data="chartData"
     :chart-id="chartId"
@@ -13,28 +13,28 @@
 </template>
 
 <script>
-import { PolarArea } from 'vue-chartjs/legacy'
+import { Radar } from '../../src/index'
 
 import {
   Chart as ChartJS,
   Title,
   Tooltip,
   Legend,
-  ArcElement,
+  PointElement,
   RadialLinearScale
 } from 'chart.js'
 
-ChartJS.register(Title, Tooltip, Legend, ArcElement, RadialLinearScale)
+ChartJS.register(Title, Tooltip, Legend, PointElement, RadialLinearScale)
 
 export default {
-  name: 'LegacyPolarArea',
+  name: 'RadarChart',
   components: {
-    PolarArea
+    Radar
   },
   props: {
     chartId: {
       type: String,
-      default: 'polar-chart'
+      default: 'radar-chart'
     },
     datasetIdKey: {
       type: String,
@@ -77,6 +77,7 @@ export default {
           {
             label: 'My First dataset',
             backgroundColor: 'rgba(179,181,198,0.2)',
+            borderColor: 'rgba(179,181,198,1)',
             pointBackgroundColor: 'rgba(179,181,198,1)',
             pointBorderColor: '#fff',
             pointHoverBackgroundColor: '#fff',
@@ -86,6 +87,7 @@ export default {
           {
             label: 'My Second dataset',
             backgroundColor: 'rgba(255,99,132,0.2)',
+            borderColor: 'rgba(255,99,132,1)',
             pointBackgroundColor: 'rgba(255,99,132,1)',
             pointBorderColor: '#fff',
             pointHoverBackgroundColor: '#fff',
